+++ fractal_tasks_core/tasks/napari_workflows_wrapper.py
@@ -20,7 +20,6 @@ from pathlib import Path
 from typing import Any
 from typing import Dict
 from typing import List
-from typing import Optional
 from typing import Sequence
 
 import anndata as ad
@@ -620,25 +619,10 @@ def napari_workflows_wrapper(
 
 
 if __name__ == "__main__":
-    from pydantic import BaseModel
-    from pydantic import Extra
-    from fractal_tasks_core._utils import run_fractal_task
-
-    class TaskArguments(BaseModel, extra=Extra.forbid):
-        input_paths: Sequence[str]
-        output_path: str
-        metadata: Dict[str, Any]
-        component: str
-        workflow_file: str
-        input_specs: Dict[str, Dict[str, str]]
-        output_specs: Dict[str, Dict[str, str]]
-        input_ROI_table: Optional[str]
-        level: Optional[int]
-        relabeling: Optional[bool]
-        expected_dimensions: Optional[int]
+    from fractal_tasks_core.tasks._utils import run_fractal_task
 
     run_fractal_task(
         task_function=napari_workflows_wrapper,
-        TaskArgsModel=TaskArguments,
+        coerce_and_validate=True,
         logger_name=logger.name,
     )